+++ packages/frontend/web/components/Header/Header.tsx
@@ -5,7 +5,7 @@ import { Logo } from './Logo';
 import { EditionDropdown } from './EditionDropdown';
 import { Links } from './Links/Links';
 import { ReaderRevenueLinks } from './ReaderRevenueLinks';
-import { Nav } from './Nav';
+import { Nav } from './Nav/Nav';
 import { getCookie } from '@frontend/web/browser/cookie';
 import { Hide } from '@frontend/web/components/Hide';
