+++ packages/frontend/web/components/Header/Nav/ExpandedMenu/ExpandedMenuToggle/ExpandedMenuToggle.tsx
