+++ packages/frontend/web/components/Header/Nav/ExpandedMenu/ExpandedMenuToggle/VeggieBurger.tsx
